Refactor: move NetworkParams and DevParams into separate module
